split script into files
